+++ NEWS.md
@@ -1,6 +1,6 @@
 # targets 0.13.5.9000
 
-
+* Move `#!/bin/sh` line to the top of SLURM `clustermq` template file (#944, #955, @GiuseppeTT).
 
 # targets 0.13.5